py_exec/requirements.txt
mysql-connector-python==8.0.28
pandas==1.4.2
